Basic loss filter
The loss filter provides basic loss tracking in a monotonically
increasing sequence of sequence numbers. It also keeps track of the
number of obviously out-of-order seqnums, as well as the total number of
samples in the period.

It currently cannot work around sequence space wrapping, and consider
late packets as not lost afterall. Some SACK-like approach should be
considered.

refs #646

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>