Handle same MS names with different schemas
The rename attribute of the <mp /> element can already be used to
explicitely rename the stream. This commit introduces some basic
mechanism in the server to automatically rename an MS if a table by the
same name already exists in the database. Doing so, it keeps looking for
a similarly renamed table with the same schema if possible.

At the moment, an arbitrary limit is set to only 9 iterations of the
renaming loop, as it would be unwise for such big experiments that use
more than 9 similar applications with different filter configuration to
not explicitely name each instances. This limit could be easily
increased, though.

fixes: #435

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>