Search MacPorts, Homebrew and Fink paths on OS X
Use config*'s host_os and host_vendor to determine whether the host is a
MacOS X machine, and try to add the default search paths for MacPorts,
Homebrew and Fink to the relvant environment variables if they exist.

This can be disabled with --without-macosx. This could be useful if the
order is not the desired one. Custom paths can still be set through
environment variables when calling configure.

fixes: 749

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>