Link OML4R Gem build to liboml2's autotools build
The gem is installed in $datarootdir/oml2. This has the added benefit
of not having to hardcode the OML version number into the Ruby code;
configure.ac has it all!

Rake (or Bundler's rules) is not quite happy with VPATHs builds
triggered this way as the files are not where it expects to find them.
We use AC_CONFIG_COMMANDS in configure to copy the needed files to the
VPATH (if srcdir is no '.'), and add an Automake conditional (on the
same condition) to remove these files on `make clean'.

Notes to self about how to do this type of file copy depending on the
build type: AC_OUTPUT([...]) is deprecated, while AC_CONFIG_LINK
unconditionally cleans up the symlinks it created; it also breaks the
dist targets. Related to this latter issue, and unexpectedly, GNU cp -f
does not imply --remove-destination as BSD's does; also, GNU test -f
doesn't fail on symlinks, though they don't look very regular to me
(hence the dist breaking)... Anyway, using AC_CONFIG_COMMANDS is cleaner
and there is no need to change the dist target command to override cp or
test's command line, which would have been painful.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>

Fix make dist with VPATH builds

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>