Basic event-hook framework
The server takes a new argument: -H, --event-hook=HOOK where HOOK is a
program which accepts commands on STDIN, and outputs some messages on
STDOUT.

The server creates two pipes, forks, redirects the child's stdin and
stdout to/from these pipes, then executes the hook program

The hook program is expected to first print an identifying banner ("OML
HOOK READY"), then wait for commands on stdin.

Though a reverse pipe is also created for the hook's stdout to be
available to the main server process, it is not currently used for
anything else than getting the banner. Also, a read(2) on it from the
main process is blocking.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>

refs: #756