Implement omlc_inject_metadata
fixes: #418

This is does the job but is a very raw and inflexible solution. We need
structural changes on both ends:
1) Add a real MS record for schema 0 at the client
2) Provide a single generic function to write data (an OmlValue(U)
array), possibly timestamped, into an MS
2.bis) Don't let filters use it, but rather get results from them, then
write
3) Refactor the _experiment_metadata table to have a primary unique id
(not 'key', which shouldn't be primary), a sender_id, timestamps.

Also, we do not currently protect against a crafty instrumentation
trying to overwrite existing metadata (solvable with (3)).

refs: #814

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>