reverseme! Workaround NULL field injection
Inject an empty string in the metadata MP if the field name is not
provided or invalid. Injection a NULL string should work, but, at least
in the SQLite3 system tests with blobgen, the mpname ends up as the
fname too, for an unclear reason.

The real problem needs to be understood and fixed, and this commit
reverted.

refs: #929
refs: #997