Refactor type handling oml2-scaffold
All declaration and mapping of various types, be they OML, C or popt(3),
are now declared once in a hash of hashes, which is then extended with
deprecated values for backward compatibility.

Also, warn when using :real or :float.

Finally, generate code showing off all the valid, currently supported,
types.

refs: #715

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>