Make BufferedWriter keep the lock when writing
BufferedWriter::processChain() tries to write the contents of an
MBuffer of the chain out to the stream. Prior to doing the writing, it
actually released the lock on the full BufferedWriter object, which
means that the chain could actually be modified, and particularly, the
MBuffer from which the function was currently writing into the stream.
Additionaly, despite not having the lock, it still did fiddle with the
MBuffer state itself.

The tests were failing due to that, and it's surprisingly lucky that
this never failed in real life. That said, having a single giant lock on
the BufferedWriter might not be the most efficient as, if the MBuffer
needs to be protected, the rest of the chain could still continue to be
used, and content at least added.

Therefore, this fix probably has performance drawbacks. A side question
is also whether the sleep(3) call in the middle of the while loop after
the call to self->writeFunc() is very wise; exiting with status 0 should
probably be enough there.

fixes: #1072 @6h, and probably many others
refs: #1104

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>