Add newwindow method to filters
This allows to separate processing and generation of output to a writer
from clearing the state to start a new window. This separation is
needed in case one MS is output into multiple writers: the same output
needs to be sent multiple time, and cleared only once, when all writers
have received the data.

fixes: #1340

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>